+++ x/foundation/keeper/grant_test.go
@@ -7,14 +7,22 @@ import (
 
 func (s *KeeperTestSuite) TestGrant() {
 	testCases := map[string]struct {
-		grantee sdk.AccAddress
-		auth    foundation.Authorization
-		valid   bool
+		malleate func(ctx sdk.Context)
+		grantee  sdk.AccAddress
+		auth     foundation.Authorization
+		valid    bool
 	}{
 		"valid authz": {
 			grantee: s.members[0],
 			auth:    &foundation.ReceiveFromTreasuryAuthorization{},
 			valid:   true,
+		},
+		"not being censored": {
+			malleate: func(ctx sdk.Context) {
+				s.keeper.SetParams(ctx, foundation.Params{})
+			},
+			grantee: s.members[0],
+			auth:    &foundation.ReceiveFromTreasuryAuthorization{},
 		},
 		"override attempt": {
 			grantee: s.stranger,
@@ -25,6 +33,9 @@ func (s *KeeperTestSuite) TestGrant() {
 	for name, tc := range testCases {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
+			if tc.malleate != nil {
+				tc.malleate(ctx)
+			}
 
 			err := s.keeper.Grant(ctx, tc.grantee, tc.auth)
 			if tc.valid {
@@ -69,15 +80,28 @@ func (s *KeeperTestSuite) TestRevoke() {
 
 func (s *KeeperTestSuite) TestAccept() {
 	testCases := map[string]struct {
-		grantee sdk.AccAddress
-		msg     sdk.Msg
-		valid   bool
+		malleate func(ctx sdk.Context)
+		grantee  sdk.AccAddress
+		msg      sdk.Msg
+		valid    bool
 	}{
 		"valid request": {
 			grantee: s.stranger,
 			msg: &foundation.MsgWithdrawFromTreasury{
 				Operator: s.operator.String(),
 				To:       s.stranger.String(),
+				Amount:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.OneInt())),
+			},
+			valid: true,
+		},
+		"not being censored": {
+			malleate: func(ctx sdk.Context) {
+				s.keeper.SetParams(ctx, foundation.Params{})
+			},
+			grantee: s.members[0],
+			msg: &foundation.MsgWithdrawFromTreasury{
+				Operator: s.operator.String(),
+				To:       s.members[0].String(),
 				Amount:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.OneInt())),
 			},
 			valid: true,
@@ -95,6 +119,9 @@ func (s *KeeperTestSuite) TestAccept() {
 	for name, tc := range testCases {
 		s.Run(name, func() {
 			ctx, _ := s.ctx.CacheContext()
+			if tc.malleate != nil {
+				tc.malleate(ctx)
+			}
 
 			err := s.keeper.Accept(ctx, tc.grantee, tc.msg)
 			if tc.valid {