+++ x/foundation/keeper/keeper.go
@@ -7,7 +7,6 @@ import (
 	"github.com/line/lbm-sdk/codec"
 	sdk "github.com/line/lbm-sdk/types"
 	"github.com/line/lbm-sdk/x/foundation"
-	"github.com/line/lbm-sdk/x/stakingplus"
 )
 
 // Keeper defines the foundation module Keeper
@@ -21,9 +20,8 @@ type Keeper struct {
 	router *baseapp.MsgServiceRouter
 
 	// keepers
-	authKeeper    foundation.AuthKeeper
-	bankKeeper    foundation.BankKeeper
-	stakingKeeper foundation.StakingKeeper
+	authKeeper foundation.AuthKeeper
+	bankKeeper foundation.BankKeeper
 
 	feeCollectorName string
 
@@ -40,7 +38,6 @@ func NewKeeper(
 	router *baseapp.MsgServiceRouter,
 	authKeeper foundation.AuthKeeper,
 	bankKeeper foundation.BankKeeper,
-	stakingKeeper foundation.StakingKeeper,
 	feeCollectorName string,
 	config foundation.Config,
 ) Keeper {
@@ -50,7 +47,6 @@ func NewKeeper(
 		router:           router,
 		authKeeper:       authKeeper,
 		bankKeeper:       bankKeeper,
-		stakingKeeper:    stakingKeeper,
 		feeCollectorName: feeCollectorName,
 		config:           config,
 	}
@@ -59,22 +55,4 @@ func NewKeeper(
 // Logger returns a module-specific logger.
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", "x/"+foundation.ModuleName)
-}
-
-// Cleaning up the states
-func (k Keeper) Cleanup(ctx sdk.Context) {
-	msgTypeURL := stakingplus.CreateValidatorAuthorization{}.MsgTypeURL()
-	var createValidatorGrantees []sdk.AccAddress
-	k.iterateAuthorizations(ctx, func(grantee sdk.AccAddress, authorization foundation.Authorization) (stop bool) {
-		if authorization.MsgTypeURL() == msgTypeURL {
-			createValidatorGrantees = append(createValidatorGrantees, grantee)
-		}
-		return false
-	})
-
-	for _, grantee := range createValidatorGrantees {
-		if err := k.Revoke(ctx, grantee, msgTypeURL); err != nil {
-			panic(err)
-		}
-	}
 }